Remove warnings

+++ src/RR32Can/Engine.cpp
@@ -20,7 +20,7 @@ void Engine::print() const {
   EngineShortInfo::print();
   if (availability == AvailabilityStatus::FULL_DETAILS) {
     printf("UID: %#10x Proto: %s Addr: %i V: %i Dir: %i, F: %i\n", uid,
-           protocol.data(), address, velocity, direction, functionBits);
+           protocol.data(), address, velocity, static_cast<uint8_t>(direction), functionBits);
   }
 }
